+++ lib/biotcm/databases.rb
@@ -1,7 +1,11 @@
-#!/usr/bin/env ruby
 # encoding: UTF-8
 
 # Namespace of all classes who interactive with databases
-# or files downloaded from databases
+# directly or indirectly
 module BioTCM::Databases
+  BioTCM.set_autoloaders(
+    {
+    },
+    self
+  )
 end