+++ test/modules/test_utility.rb
@@ -1,0 +1,6 @@
+# encoding: UTF-8
+require 'test_helper'
+
+class BioTCM_Modules_Utility_Test < Test::Unit::TestCase
+  #TODO: add tests
+end