+++ test/test_helper.rb
@@ -1,0 +1,4 @@
+# encoding: UTF-8
+require 'test/unit'
+require 'shoulda-context'
+require 'biotcm'