+++ HISTORY.md
@@ -2,8 +2,9 @@ History
 ===============
 
 ### By now 2014-03-14
-* 1 major enhancement
+* 2 major enhancement
 	* Completed the design of architectures
+	* Added BioTCM::Modules::Utility
 
 ### 0.0.0 2014-03-06
 * 1 major enhancement